CDH-24889: Sync search-mr with upstream solr-4.10.3

+++ search-mr/src/test/java/org/apache/solr/hadoop/MRUnitBase.java
@@ -17,32 +17,45 @@
 package org.apache.solr.hadoop;
 
 import java.io.File;
+import java.io.IOException;
+import java.util.Locale;
 
+import org.apache.commons.io.FileUtils;
 import org.apache.hadoop.conf.Configuration;
+import org.apache.solr.SolrTestCaseJ4;
 import org.apache.solr.hadoop.morphline.MorphlineMapRunner;
 import org.junit.AfterClass;
-import org.junit.Assert;
 import org.junit.BeforeClass;
 
-public class MRUnitBase extends Assert {
+public abstract class MRUnitBase extends SolrTestCaseJ4 {
   
-  protected static final String RESOURCES_DIR = "target/test-classes";
+  protected static final String RESOURCES_DIR = getFile("morphlines-core.marker").getParent();
   protected static final String DOCUMENTS_DIR = RESOURCES_DIR + "/test-documents";
   protected static File solrHomeZip;
 
   @BeforeClass
   public static void setupClass() throws Exception {
+    assumeFalse("This test fails on UNIX with Turkish default locale (https://issues.apache.org/jira/browse/SOLR-6387)",
+        new Locale("tr").getLanguage().equals(Locale.getDefault().getLanguage()));
     solrHomeZip = SolrOutputFormat.createSolrHomeZip(new File(RESOURCES_DIR + "/solr/mrunit"));
     assertNotNull(solrHomeZip);
   }
 
   @AfterClass
   public static void teardownClass() throws Exception {
-    solrHomeZip.delete();
+    if (solrHomeZip != null) solrHomeZip.delete();
   }
   
-  protected void setupHadoopConfig(Configuration config) {
+  protected void setupHadoopConfig(Configuration config) throws IOException {
+    
+    String tempDir = createTempDir().getAbsolutePath();
+
+    FileUtils.copyFile(new File(RESOURCES_DIR + "/custom-mimetypes.xml"), new File(tempDir + "/custom-mimetypes.xml"));
+
+    UtilsForTests.setupMorphline(tempDir, "test-morphlines/solrCellDocumentTypes", true, RESOURCES_DIR);
+    
+    config.set(MorphlineMapRunner.MORPHLINE_FILE_PARAM, tempDir + "/test-morphlines/solrCellDocumentTypes.conf");
     config.set(SolrOutputFormat.ZIP_NAME, solrHomeZip.getName());
-    config.set(MorphlineMapRunner.MORPHLINE_FILE_PARAM, RESOURCES_DIR + "/test-morphlines/solrCellDocumentTypes.conf");
   }
+  
 }
